+++ vscode-extension/src/lib/functions/getAlObjectEntityIds.ts
@@ -1,0 +1,5 @@
+import { ALObject, ALUniqueEntity } from "@vjeko.com/al-parser-types-ninja";
+
+export function getAlObjectEntityIds(object: ALObject): ALUniqueEntity[] {
+    return object.fields || object.values || [];
+}