+++ vscode-extension/src/lib/types/AssignedALObject.ts
@@ -4,4 +4,7 @@ import { ALObjectType } from "./ALObjectType";
 export interface AssignedALObject {
     type: ALObjectType;
     id: number;
+    possiblePaths?: string[];
+    name?: string;
+    fieldOrValueIds?: number[];
 }